Moved PHP-DI-related classes to its own namespace
