+++ src/Provider/StandardServicesProvider.php
@@ -7,7 +7,7 @@ namespace Noctis\KickStart\Provider;
 use Laminas\Diactoros\StreamFactory;
 use Noctis\KickStart\Http\Response\Attachment\AttachmentFactory;
 use Noctis\KickStart\Http\Response\Attachment\AttachmentFactoryInterface;
-use Noctis\KickStart\Service\Container\PhpDiContainer;
+use Noctis\KickStart\Service\Container\PhpDi\Container;
 use Noctis\KickStart\Service\Container\SettableContainerInterface;
 use Noctis\KickStart\Service\TemplateRendererInterface;
 use Noctis\KickStart\Service\TwigRenderer;
@@ -23,8 +23,8 @@ final class StandardServicesProvider implements ServicesProviderInterface
     {
         return [
             AttachmentFactoryInterface::class => AttachmentFactory::class,
-            ContainerInterface::class => PhpDiContainer::class,
-            SettableContainerInterface::class => PhpDiContainer::class,
+            ContainerInterface::class => Container::class,
+            SettableContainerInterface::class => Container::class,
             StreamFactoryInterface::class => StreamFactory::class,
             TemplateRendererInterface::class  => TwigRenderer::class,
         ];
