+++ src/Service/Container/ContainerBuilderInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Noctis\KickStart;
+namespace Noctis\KickStart\Service\Container;
 
 use Noctis\KickStart\Provider\ServicesProviderInterface;
 use Psr\Container\ContainerInterface;
